+++ packages/camera/camera/example/lib/main.dart
@@ -623,7 +623,9 @@ class _CameraExampleHomeState extends State<CameraExampleHome>
   }
 
   void onNewCameraSelected(CameraDescription cameraDescription) async {
-    final previousCameraController = controller;
+    if (controller != null) {
+      await controller!.dispose();
+    }
 
     final CameraController cameraController = CameraController(
       cameraDescription,
@@ -633,10 +635,6 @@ class _CameraExampleHomeState extends State<CameraExampleHome>
     );
 
     controller = cameraController;
-
-    if (mounted) {
-      setState(() {});
-    }
 
     // If the controller is updated then update the UI.
     cameraController.addListener(() {